phase1046c-revived-wae-site-fixes-wave2(engine+samples+.clang-tidy)
Wave 2 of the revived-family WAE cleanup: 48 sites across 17 files
(chunk-1 recheck surfaced log/io/profile/ui sites the first sweep's
interleaved parallel output had masked + chunk-2 render-tree sites).
Fixed via 3 parallel developer agents on disjoint file sets + 2
orchestrator-owned files; verified centrally.

  - 16x lock_guard -> scoped_lock (AuditTrail / ConsoleLogger /
    JsonLogger / RingBufferSink).
  - 5x escaped strings -> raw string literals (JsonLogger fragments
    + test_log; content byte-identical).
  - 3x rewind(f) -> (void)std::fseek(f, 0, SEEK_SET)
    (bugprone-unsafe-functions).
  - 10x implicit-widening-of-multiplication (test_io / test_mem /
    test_denoise x8): size_t cast BEFORE the multiply.
  - 2x misplaced-widening-cast (Denoise.hpp: cast the operand, not
    the sum; dy/dx in [-2,2] so indices identical).
  - 2x cert-flp30-c float-induction loops -> int induction
    (test_brdf_sss; identical sample points verified).
  - 2x return-braced-init-list (Input.hpp / Widgets.hpp span
    returns); 3x use-auto (Framing / hello_ui / test_restir_di);
    1x use-ranges (StatsAggregator sort); 1x faster-string-find;
    2x isolate-declaration; 1x inefficient-vector-operation
    (reserve); 1x confusable-identifiers (Sss.hpp exp1/exp3 ->
    exp_a/exp_b, locals only); 1x misleading-indentation (Ddgi.hpp
    whitespace-only); 2x member-init NSDMI (DeferredTlas +
    DeferredBuffer destroy_at_frame); 2x unused namespace aliases
    (hello_ui).
  - 2x judged NOLINTNEXTLINE with rationale: _CRT_SECURE_NO_WARNINGS
    (documented MSVC opt-out needs the reserved name) +
    PanicException baseclass (deriving std::exception would break
    aggregate init at the throw site; test intent preserved).

.clang-tidy: 3 alias-disables added mirroring already-disabled
primaries that the revived hicpp/cert families re-exposed:
hicpp-braces-around-statements, hicpp-explicit-conversions,
cert-msc32-c (sibling of disabled cert-msc51-cpp).

Build: ninja-debug clean.
Tests: 261/261 PASS.

+++ engine/foundation/io/tests/test_io.cpp
@@ -101,7 +101,7 @@ TEST(BinaryStream, SeekAndPosition)
         w.write<std::uint32_t>(i * 7);
 
     cd::io::BinaryReader r { w.data() };
-    r.seek(4 * 3);  // skip first 3
+    r.seek(static_cast<std::size_t>(4) * 3);  // skip first 3
     EXPECT_EQ(r.position(), 12u);
     EXPECT_EQ(*r.read<std::uint32_t>(), 3u * 7u);
